+++ CHANGELOG.md
@@ -6,9 +6,11 @@ View all releases at: <https://github.com/trimble-oss/modus-web-components/relea
 
 ## Unreleased
 
-### Fixed
+### Added
 
-- Modus Accordions now have transition animations
+- Added password text toggle feature to `modus-text-input` component.
+
+- Accordions now have transition animations
 
 ## 0.1.28 - 2022-12-06